requirements.txt
asgiref==3.2.7
celery==4.4.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.6
django-cors-headers==3.3.0
django-extensions==2.2.9
django-celery-beat==2.0.0
dj-database-url
django-import-export==2.2.0
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
django-sslserver
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
et-xmlfile==1.0.1
gunicorn
idna==2.9
inflection==0.4.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.3
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
tablib
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0
